feat: extract ASP.NET Core middleware into new Vcr.HttpRecorder.AspNetCore project
